requirements.txt
flask
gunicorn
python-docx
PyPDF2